@KristofferC, could you do a speed test on your machine, this seems to give me a > 2x first run  speed up but 10% drop in later runs. Acceptable?