@KristofferC 2017-09-21
First parse from 9.4s -> 3.9s. Benchmark from 26.6ms -> 31.6 ms. I think for now this is probably worth it in terms of user experience.